plc-modbus-sim/requirements.txt
pymodbus==3.7.4
psycopg2-binary==2.9.10
